Update name, version number, dependencies
Changed paho dependency to new version 1.2, reverted component version
and GeoEvent SDK dependency to 10.4.1 (but also testing the same build
with 10.5.1). Also changed transport name to just MQTT for both inbound
and outbound, as this is more consistent with built-in transports in the
GeoEvent Manager UI. This does break any references to the previous
names e.g. MqttInboundTransport/10.4.0, but they are broken already by
the version number change.